+++ services/src/main/java/org/keycloak/protocol/oidc/mappers/HardcodedClaimBasedOnAttributeMapper.java
@@ -78,8 +78,9 @@ public class HardcodedClaimBasedOnAttributeMapper extends AbstractOIDCProtocolMa
             Collection<String> userAttributeValues = KeycloakModelUtils.resolveAttribute(userSession.getUser(), userAttribute , false);
 
             if (userAttributeValues.stream().anyMatch(possibleAttributeValues::contains)) {
-                List<String> attributeValues = Arrays.asList(mappingModel.getConfig().get(ProtocolMapperUtils.CLAIM_VALUE).split(Constants.CFG_DELIMITER));
-                OIDCAttributeMapperHelper.mapClaim(token, mappingModel, attributeValues.size() ==1 ? attributeValues.get(0) : attributeValues);
+                String attributeValue = mappingModel.getConfig().get(ProtocolMapperUtils.CLAIM_VALUE);
+                if (attributeValue == null) return;
+                OIDCAttributeMapperHelper.mapClaim(token, mappingModel, attributeValue);
             }
     }