+++ services/src/main/java/org/keycloak/protocol/oidc/mappers/HardcodedClaimBasedOnIdPMapper.java
@@ -77,8 +77,9 @@ public class HardcodedClaimBasedOnIdPMapper extends AbstractOIDCProtocolMapper i
             String userIdP = userSession.getNote(Details.IDENTITY_PROVIDER);
 
             if (userIdP != null && idpsAlias.contains(userIdP)) {
-                List<String> attributeValues = Arrays.asList(mappingModel.getConfig().get(ProtocolMapperUtils.CLAIM_VALUE).split(Constants.CFG_DELIMITER));
-                OIDCAttributeMapperHelper.mapClaim(token, mappingModel, attributeValues.size() ==1 ? attributeValues.get(0) : attributeValues);
+                String attributeValue = mappingModel.getConfig().get(ProtocolMapperUtils.CLAIM_VALUE);
+                if (attributeValue == null) return;
+                OIDCAttributeMapperHelper.mapClaim(token, mappingModel, attributeValue);
             }
 
     }